requirements/base.txt
Django==3.0.7
django-environ==0.4.5
psycopg2-binary==2.8.4
django-braces==1.14.0
pypugjs==5.9.4
django-crispy-forms==1.9.0
django-datatables-view==1.19.1
social-auth-app-django==4.0.0
channels==2.4.0
channels-redis==2.4.2
django-cleanup==5.0.0
django-markdownx==3.0.1
django-crispy-forms==1.7.0
Markdown==3.2.2
djangorestframework==3.11.1
